+++ src/prime_rl/inference/config.py
@@ -1,4 +1,3 @@
-import os
 from argparse import Namespace
 from typing import Annotated, Any, Literal
 
@@ -162,6 +161,13 @@ class InferenceConfig(BaseSettings):
         ),
     ] = None
 
+    enable_prefix_caching: Annotated[
+        bool | None,
+        Field(
+            description="Whether to enable prefix caching. Passed to vLLM as `--enable-prefix-caching`",
+        ),
+    ] = None
+
     gpu_memory_utilization: Annotated[
         float,
         Field(
@@ -187,12 +193,6 @@ class InferenceConfig(BaseSettings):
     weight_broadcast: Annotated[WeightBroadcastConfig, Field(description="The weight broadcast config.")] = (
         WeightBroadcastConfig()
     )
-
-    @model_validator(mode="after")
-    def auto_setup_dynamic_lora_updating(self):
-        if self.enable_lora:
-            os.environ["VLLM_ALLOW_RUNTIME_LORA_UPDATING"] = "True"
-        return self
 
     @model_validator(mode="after")
     def round_up_max_lora_rank(self):
@@ -244,6 +244,7 @@ class InferenceConfig(BaseSettings):
             "parallel.tp": "tensor_parallel_size",
             "parallel.dp": "data_parallel_size",
             "enable_lora": "enable_lora",
+            "enable_prefix_caching": "enable_prefix_caching",
             "max_loras": "max_loras",
             "max_cpu_loras": "max_cpu_loras",
             "max_lora_rank": "max_lora_rank",